Upgrade '@emotion', '@babel' and '@floating-ui' dependencies, also update '@mui' and 'prop-types'
This commit introduces and updates a number of dependencies in the 'yarn.lock' file. It includes a significant addition of dependencies linked to '@emotion', '@babel', and '@floating-ui'. It also involves updating certain versions of '@mui' and 'prop-types'. These upgrades contribute to enhancing the stability, performance, and compatibility of the project with different packages.